Fix phinode rename bug during domsort
Since the edge list of a phi may be incomplete with respect to
predecessors, it is not safe to replace a one-value phi by its
value, unless we know that there is only one predecessor. Remove
the incorrect logic from domsort. We do also do the correct
transformation in compaction right afterwards these days.